requirements.txt
albumentations==1.1.0
data==0.4
deepsparse==1.3.1
deepsparse_nightly==1.2.0.20221024
detectron2==0.6
faiss==1.7.1
fontTools==4.38.0
kornia==0.6.4
layoutparser==0.3.4
matplotlib==3.5.1
mmcv==1.7.1
mmcv_full==1.3.17
mmdet==2.27.0
mmdet.egg==info
nltk==3.7
numpy==1.22.4
omegaconf==2.2.3
onnx==1.10.1
onnxruntime==1.12.1
opencv_python_headless==4.5.5.64
pandas==1.3.4
Pillow==9.4.0
protobuf==4.21.12
pycocotools==2.0.4
pytesseract==0.3.9
pytorch_metric_learning==1.4.0
PyYAML==6.0
requests==2.26.0
scikit_image==0.19.1
scipy==1.7.2
segmentation_models_pytorch==0.2.1
setuptools==58.0.4
shapely==2.0.0
skimage==0.0
sparseml==1.3.1
sparseml_nightly==1.3.0.20221108
sparsezoo==1.2.0
symspellpy==6.7.6
timm==0.6.2.dev0
torch==1.10.0
torchmetrics==0.7.2
torchvision==0.11.1
tqdm==4.64.1
transformers==4.18.0
vit_pytorch==0.25.1
yolov5==7.0.7
